+++ app/src/main/java/com/example/fglearning/repository/FlashcardRepository.kt
@@ -6,9 +6,9 @@ import com.example.fglearning.database.entity.Flashcard
 import com.example.fglearning.database.entity.InsertLetter
 
 class FlashcardRepository(private val flashcardDao: FlashcardDao) {
-    suspend fun getById(id: Int): Flashcard? = flashcardDao.getById(id)
+    suspend fun getById(id: Long): Flashcard? = flashcardDao.getById(id)
 
-    suspend fun deleteById(id: Int) = flashcardDao.deleteById(id)
+    suspend fun deleteById(id: Long) = flashcardDao.deleteById(id)
 
     suspend fun insert(flashcard: Flashcard) = flashcardDao.insert(flashcard)
